Bug 98614: Add localconfig option to override the submission SSL cert
Right now Zimbra doesn't support virtual hosts for authenticated SMTP.
And in the case of SNI never will as long as nginx isn't used as a SMTP
proxy as well (cf. BZ-75623).

Even the simplest form isn't possible where the hosts use the ZCS CA
managed certificates internally but show another name like mail.example.com
to the outside world.  Or, where you've got two MTAs/MXes with correct
certs for their internal hostnames but provide a DNS alias like the above
for authenticated SMTP which fails over to the other MTA.

A workaround in 8.0 was to override the localconfig variables
postfix_smtpd_tls_cert_file and postfix_smtpd_tls_key_file and make them
point to a custom SSL certificate.  With the ldapification in 8.6 this
feature went away.

This change introduces the two localconfig variables
* postfix_submission_smtpd_tls_key_file
* postfix_submission_smtpd_tls_cert_file
which allow to override the certs of the submissions ports (587 and 465)
with another local, probably domain, certificate file.  We use this patch
widely in production.  It is intended as a workaround until virtual hosts
are supported for port 587 and 465.